Revert "[Mold API, UI, Agent] europa 브랜치 커스터마이징 작업"

+++ engine/storage/volume/src/main/java/org/apache/cloudstack/storage/volume/VolumeServiceImpl.java
@@ -328,11 +328,6 @@ public class VolumeServiceImpl implements VolumeService {
         } else {
             vo.processEvent(Event.OperationFailed);
             errMsg = result.getResult();
-            VolumeVO volume = volDao.findById(vo.getId());
-            if (volume != null && volume.getState() == State.Allocated && volume.getPodId() != null) {
-                volume.setPoolId(null);
-                volDao.update(volume.getId(), volume);
-            }
         }
         VolumeApiResult volResult = new VolumeApiResult((VolumeObject)vo);
         if (errMsg != null) {
@@ -512,9 +507,6 @@ public class VolumeServiceImpl implements VolumeService {
                     long storagePoolId = snapStoreVo.getDataStoreId();
                     StoragePoolVO storagePoolVO = storagePoolDao.findById(storagePoolId);
 
-                    if (StoragePoolType.StorPool.equals(storagePoolVO.getPoolType())) {
-                        continue;
-                    }
                     if (storagePoolVO.isManaged()) {
                         DataStore primaryDataStore = dataStoreMgr.getPrimaryDataStore(storagePoolId);
                         Map<String, String> mapCapabilities = primaryDataStore.getDriver().getCapabilities();
@@ -525,10 +517,10 @@ public class VolumeServiceImpl implements VolumeService {
                         if (!supportsStorageSystemSnapshots) {
                             _snapshotStoreDao.remove(snapStoreVo.getId());
                         }
-                    } else if (HypervisorType.KVM.equals(vo.getHypervisorType())) {
-                        deleteKvmSnapshotOnPrimary(snapStoreVo);
                     } else {
-                        _snapshotStoreDao.remove(snapStoreVo.getId());
+                        if (!StoragePoolType.StorPool.equals(storagePoolVO.getPoolType())) {
+                            _snapshotStoreDao.remove(snapStoreVo.getId());
+                        }
                     }
                 }
                 snapshotApiService.markVolumeSnapshotsAsDestroyed(vo);
@@ -541,21 +533,6 @@ public class VolumeServiceImpl implements VolumeService {
         }
         context.getFuture().complete(apiResult);
         return null;
-    }
-
-    /**
-     * Deletes the snapshot from primary storage if the only storage associated with the snapshot is of the Primary role; else, just removes the primary record on the DB.
-     * */
-    protected void deleteKvmSnapshotOnPrimary(SnapshotDataStoreVO snapshotDataStoreVO) {
-        List<SnapshotDataStoreVO> snapshotDataStoreVOList = _snapshotStoreDao.findBySnapshotId(snapshotDataStoreVO.getSnapshotId());
-        for (SnapshotDataStoreVO snapshotStore : snapshotDataStoreVOList) {
-            if (DataStoreRole.Image.equals(snapshotStore.getRole())) {
-                _snapshotStoreDao.remove(snapshotDataStoreVO.getId());
-                return;
-            }
-        }
-
-        snapshotApiService.deleteSnapshot(snapshotDataStoreVO.getSnapshotId(), null);
     }
 
     @Override
@@ -1072,7 +1049,7 @@ public class VolumeServiceImpl implements VolumeService {
             try {
                 grantAccess(templateOnPrimary, destHost, destPrimaryDataStore);
             } catch (Exception e) {
-                throw new StorageAccessException(String.format("Unable to grant access to template: %s on host: %s", templateOnPrimary.getImage(), destHost), e);
+                throw new StorageAccessException(String.format("Unable to grant access to template: %s on host: %s", templateOnPrimary.getImage(), destHost));
             }
 
             templateOnPrimary.processEvent(Event.CopyingRequested);
@@ -1194,7 +1171,7 @@ public class VolumeServiceImpl implements VolumeService {
             try {
                 grantAccess(srcTemplateOnPrimary, destHost, destPrimaryDataStore);
             } catch (Exception e) {
-                throw new StorageAccessException(String.format("Unable to grant access to src template: %s on host: %s", srcTemplateOnPrimary, destHost), e);
+                throw new StorageAccessException(String.format("Unable to grant access to src template: %s on host: %s", srcTemplateOnPrimary, destHost));
             }
 
             _volumeDetailsDao.addDetail(volumeInfo.getId(), volumeDetailKey, String.valueOf(templatePoolRef.getId()), false);
@@ -1271,10 +1248,6 @@ public class VolumeServiceImpl implements VolumeService {
         }
 
         if (volume.getState() == State.Allocated) { // Possible states here: Allocated, Ready & Creating
-            if (volume.getPodId() != null) {
-                volume.setPoolId(null);
-                volDao.update(volume.getId(), volume);
-            }
             return;
         }
 
@@ -1445,7 +1418,7 @@ public class VolumeServiceImpl implements VolumeService {
                 try {
                     grantAccess(templateOnPrimary, destHost, destPrimaryDataStore);
                 } catch (Exception e) {
-                    throw new StorageAccessException(String.format("Unable to grant access to template: %s on host: %s", templateOnPrimary, destHost), e);
+                    throw new StorageAccessException(String.format("Unable to grant access to template: %s on host: %s", templateOnPrimary, destHost));
                 }
 
                 templateOnPrimary.processEvent(Event.CopyingRequested);
@@ -2537,7 +2510,7 @@ public class VolumeServiceImpl implements VolumeService {
         try {
             volume.processEvent(Event.ResizeRequested);
         } catch (Exception e) {
-            logger.debug("Failed to change volume state to resize", e);
+            logger.debug("Failed to change state to resize", e);
             result.setResult(e.toString());
             future.complete(result);
             return future;
@@ -2549,8 +2522,10 @@ public class VolumeServiceImpl implements VolumeService {
         try {
             volume.getDataStore().getDriver().resize(volume, caller);
         } catch (Exception e) {
-            logger.debug("Failed to resize volume", e);
+            logger.debug("Failed to change state to resize", e);
+
             result.setResult(e.toString());
+
             future.complete(result);
         }
 
@@ -2594,7 +2569,7 @@ public class VolumeServiceImpl implements VolumeService {
             try {
                 volume.processEvent(Event.OperationFailed);
             } catch (Exception e) {
-                logger.debug("Failed to change volume state (after resize failure)", e);
+                logger.debug("Failed to change state", e);
             }
             VolumeApiResult res = new VolumeApiResult(volume);
             res.setResult(result.getResult());
@@ -2605,8 +2580,13 @@ public class VolumeServiceImpl implements VolumeService {
         try {
             volume.processEvent(Event.OperationSuccessed);
         } catch (Exception e) {
-            logger.debug("Failed to change volume state (after resize success)", e);
+            logger.debug("Failed to change state", e);
+            VolumeApiResult res = new VolumeApiResult(volume);
+            res.setResult(result.getResult());
+            future.complete(res);
+            return null;
         }
+
         VolumeApiResult res = new VolumeApiResult(volume);
         future.complete(res);
